Create practiceq1
Q1. Write a program to create a 2D array (say new2D) that contains 2 rows. The first row
should store an array of integers and the second row should store an array of float values.
Take the number of elements in the first array and the second array as input from command
line arguments. Then fill the 2D array with values taken from the user as input using scanf.
Then Print the 2D array.


[Hint:
Create a double-pointer of the type void (with the name new2D). Then allocate memory to it
for size 2. This creates an array of size 2, where each location can store address of another
variable (or an array). Then use appropriate typecasting to allocate an integer array in the first
location of new2D and a float array in the second location of new2D, using the respective sizes
taken from the user as input. 
END OF HINT]